✨ feat: refactor the handleUpload to restrict with xlsx
