+++ src/pages/Hydro/Sider.jsx
@@ -9,23 +9,12 @@ import { useState } from 'react';
 import { basename } from '@tauri-apps/api/path';
 import Papa from 'papaparse';
 import { setHydro } from '../../store/dataSlice'; 
+import { handleUpload } from '../utils/utils';
 
 
 export default function Sider({form, tabsName, handleSubmit, datas, file={name: ''}, setFile}) {
 
   const [orbit, setOrbit] = useState(false)
-
-  const handleUpload = async (id) => {
-    const filePath = await open({ multiple: false })
-    if (filePath) {
-      const filename = await basename(filePath)
-      setFile(() => ({name: filename, path: filePath}))
-      Message.success(`${filename}上传成功！`)
-      setOrbit((prev) => !prev)
-    } else {
-      Message.info('文件上传失败,请重新上传')
-    }
-  }
 
   const handleCancel = (id) => {
     if (id === 1) {
